Changed formatting of script

+++ windows-10/setup/agent.ps1
@@ -6,25 +6,23 @@ $installer = "VMware-Horizon-Agent-x86_64-2012-8.1.0-17352461.exe"
 $listConfig = "/s /v ""/qn REBOOT=ReallySuppress ADDLOCAL=Core,NGVC,RTAV,ClientDriveRedirection,V4V,VmwVaudio,PerfTracker"""
 
 # Verify connectivity
-Test-Connection $webserver -Count 1
+$connTestResult = Test-NetConnection -Computername $webserver -Port 80
+if ($connTestResult.TcpTestSucceeded){
+  # Get Horizon Agent
+  Invoke-WebRequest -Uri ($url + "/" + $installer) -OutFile C:\$installer
 
-# Get Horizon Agent
-Invoke-WebRequest -Uri ($url + "/" + $installer) -OutFile C:\$installer
+  # Unblock installer
+  Unblock-File C:\$installer -Confirm:$false -ErrorAction Stop
 
-# Unblock installer
-Unblock-File C:\$installer -Confirm:$false -ErrorAction Stop
+  # Install Horizon Agent
+  Try {
+    Start-Process C:\$installer -ArgumentList $listConfig -PassThru -Wait -ErrorAction Stop
+  } Catch {
+    Write-Error "Failed to install the Horizon Agent"
+    Write-Error $_.Exception
+    Exit -1
+  }
 
-# Install Horizon Agent
-Try 
-{
-   Start-Process C:\$installer -ArgumentList $listConfig -PassThru -Wait -ErrorAction Stop
+  # Cleanup on aisle 4...
+  Remove-Item C:\$installer -Confirm:$false
 }
-Catch
-{
-   Write-Error "Failed to install the Horizon Agent"
-   Write-Error $_.Exception
-   Exit -1 
-}
-
-# Cleanup on aisle 4...
-Remove-Item C:\$installer -Confirm:$false
